requirements.txt
cmake==3.25.2
dlib==19.24.0
face-recognition==1.3.0
keras==2.11.0
matplotlib==3.7.0
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
requests==2.28.2
scipy==1.10.1
tensorflow==2.11.0
insightface==0.7.2